requirements.txt
# Project Trishul - Python Dependencies
# Install with: pip install -r requirements.txt

# Core Dependencies
requests>=2.28.0
flask>=2.3.0
pandas>=2.0.0

# Dashboard
streamlit>=1.28.0

# Environment Variables
python-dotenv>=1.0.0

# CLI Enhancement
rich>=13.0.0
tqdm>=4.65.0

# YAML Config Support
pyyaml>=6.0

# OSINT Dependencies
dnspython>=2.4.0  # For DNS analysis
